src/main/java/com/niveksys/recipeapp/converter/CategoryCommandToCategory.java
package com.niveksys.recipeapp.converter;

import com.niveksys.recipeapp.command.CategoryCommand;
import com.niveksys.recipeapp.model.Category;

import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

import lombok.Synchronized;

@Component
public class CategoryCommandToCategory implements Converter<CategoryCommand, Category> {

    @Synchronized
    @Nullable
    @Override
    public Category convert(CategoryCommand source) {
        if (source == null) {
            return null;
        }

        final Category category = new Category();
        category.setId(source.getId());
        category.setDescription(source.getDescription());
        return category;
    }

}
